+++ src/lib/plugins/index.js
@@ -1,6 +1,7 @@
 import lintAtRules from './lintAtRules'
 import evaluateEuropaFunctions from './evaluateEuropaFunctions'
 import substituteIfAtRules from './substituteIfAtRules'
+import substituteColorAtRules from './substituteColorAtRules'
 import substituteEuropaAtRules from './substituteEuropaAtRules'
 import substituteIterateAtRules from './substituteIterateAtRules'
 import substituteUnpackAtRules from './substituteUnpackAtRules'
@@ -21,6 +22,7 @@ export default [
   lintAtRules,
   substituteEuropaAtRules,
   substituteIfAtRules,
+  substituteColorAtRules,
   evaluateEuropaFunctions,
   substituteIterateAtRules,
   substituteUnpackAtRules,