+++ stubs/defaultConfig.js
@@ -47,7 +47,8 @@ module.exports = {
     colors: {
       transparent: 'transparent',
       black: '#000',
-      white: '#fff'
+      white: '#fff',
+      grid: '#0095ff33'
     },
 
     arrows: {